fix: do not sure tooltips for issues not in the db
This prevents 'exploring' other repos with issuelinks and polluting the
database. Thus the issues need either to be imported or added via a
webhook.

We may want to change that in the future, since there could be a usecase
for single issues from other public repositories which are only
updated on displaying the tooltip.