🐛 Fix jira webhook not correctly created
There was an error that the webhook at Jira was not created when
clicking the red cross for a project in the admin section. It was
however created when clicking the green checkmark that appeared after
clicking the red cross.

The root cause is that buildAttributes with skipempty=true would only skip exactly
empty strings, but not null, false, etc. This caused multiple problems,
e.g. $INPUT->has() considered the parameter to be present, even though
it was just an empty string, etc.

Setting the $hookID to an empty string should prevent it on appearing on
projects which have no webook.

This hopefully improves #3